refactor: move to yarn workspaces (#1)
* refactor: use workspaces to enable testing of bumped versions

* test: add build & test workflow for web-app and bank2ynab-converter

* test: add results step to not need manual update of required checks when changing the matrix job
