+++ .github/workflows/test-web-app.yml
@@ -1,4 +1,4 @@
-name: Test parsers
+name: Test web-app
 
 on: [pull_request]
 
@@ -20,20 +20,17 @@ jobs:
         uses: actions/setup-node@v3.6.0
         with:
           node-version: ${{ matrix.node-version }}
+
       - name: Install, Build, and Test
         run: |
-          cd packages/ynap-parsers
+          cd packages/ynap-web-app
           yarn install
           yarn build
-          yarn test --coverage
         env:
           CI: true
 
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v3.1.3
-
   # See https://github.com/orgs/community/discussions/26822#discussioncomment-5122101
-  results-parsers:
+  results-web-app:
     if: ${{ always() }}
     runs-on: ubuntu-latest
     needs: [build]
