+++ package.json
@@ -2,6 +2,8 @@
   "name": "you-need-a-parser",
   "private": true,
   "devDependencies": {
-    "lerna": "3.22.1"
-  }
+    "lerna": "3.22.1",
+    "typescript": "4.9.5"
+  },
+  "workspaces": ["packages/*"]
 }
