+++ packages/ynap-bank2ynab-converter/package.json
@@ -16,8 +16,7 @@
   "devDependencies": {
     "@types/node": "18.16.16",
     "@types/node-fetch": "2.6.4",
-    "ts-node": "10.9.1",
-    "typescript": "3.9.10"
+    "ts-node": "10.9.1"
   },
   "dependencies": {
     "commander": "^3.0.0",
