+++ packages/ynap-parsers/package.json
@@ -26,7 +26,6 @@
     "fast-glob": "3.2.12",
     "jest": "24.9.0",
     "ts-jest": "24.3.0",
-    "typescript": "3.9.10",
     "ynap-bank2ynab-converter": "^1.14.0"
   },
   "scripts": {
