+++ packages/ynap-web-app/package.json
@@ -4,25 +4,24 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@types/jest": "24.9.1",
-    "@types/node": "18.16.16",
-    "@types/react": "16.14.42",
-    "@types/react-dom": "16.9.19",
+    "@types/node": "18.16.15",
+    "@types/jest": "29.5.1",
+    "@types/react": "18.2.6",
+    "@types/react-dom": "18.2.4",
     "file-saver": "2.0.5",
-    "gatsby": "2.32.13",
-    "gatsby-plugin-netlify": "2.11.1",
-    "gatsby-plugin-offline": "3.10.2",
-    "gatsby-plugin-react-helmet": "3.10.0",
-    "gatsby-plugin-styled-components": "3.10.0",
-    "gatsby-plugin-typescript": "2.12.1",
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "react-helmet": "5.2.1",
+    "gatsby": "5.10.0",
+    "gatsby-plugin-netlify": "5.1.0",
+    "gatsby-plugin-offline": "6.10.0",
+    "gatsby-plugin-react-helmet": "6.10.0",
+    "gatsby-plugin-styled-components": "6.10.0",
+    "gatsby-plugin-typescript": "5.10.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-helmet": "6.1.0",
     "react-helmet-async": "1.3.0",
-    "react-toastify": "5.5.0",
-    "styled-components": "4.4.1",
+    "react-toastify": "9.1.3",
+    "styled-components": "5.3.10",
     "ts-node": "10.9.1",
-    "typescript": "3.9.10",
     "ynap-parsers": "^1.15.0"
   },
   "scripts": {
@@ -42,14 +41,14 @@
   "devDependencies": {
     "@types/file-saver": "2.0.5",
     "@types/node-fetch": "2.6.4",
-    "@types/react-helmet": "5.0.21",
-    "@types/react-helmet-async": "1.0.1",
-    "@types/react-toastify": "4.0.2",
-    "@types/styled-components": "4.4.3",
-    "babel-plugin-styled-components": "1.13.3",
+    "@types/react-helmet": "6.1.6",
+    "@types/react-helmet-async": "1.0.3",
+    "@types/react-toastify": "4.1.0",
+    "@types/styled-components": "5.1.26",
+    "babel-plugin-styled-components": "2.1.3",
     "date-fns": "2.30.0",
     "encoding": "0.1.13",
-    "jest": "24.9.0",
+    "jest": "29.5.0",
     "node-fetch": "2.6.11",
     "ts-jest": "24.3.0",
     "tslint": "5.20.1"
