+++ packages/ynap-web-app/src/pages/supported-formats.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import '../styles/index.css';
 
-import { parsers, countries } from 'ynap-parsers';
+import { parsers, countries } from 'ynap-parsers/src';
 import countryNames from '../util/countries';
 import MetaTags from '../components/meta-tags';
 import { Link } from 'gatsby';
